chore: public processor commits/clears public contracts cache for tx rather than public tx simulator
