+++ yarn-project/simulator/src/public/public_processor/public_processor.test.ts
@@ -1,14 +1,19 @@
+import { CONTRACT_CLASS_PUBLISHED_MAGIC_VALUE, CONTRACT_CLASS_REGISTRY_CONTRACT_ADDRESS } from '@aztec/constants';
 import { timesParallel } from '@aztec/foundation/collection';
 import { Fr } from '@aztec/foundation/fields';
 import { sleep } from '@aztec/foundation/sleep';
 import { TestDateProvider } from '@aztec/foundation/timer';
 import { computeFeePayerBalanceLeafSlot } from '@aztec/protocol-contracts/fee-juice';
+import { bufferAsFields } from '@aztec/stdlib/abi';
 import { AvmCircuitInputs, PublicDataWrite, RevertCode } from '@aztec/stdlib/avm';
 import { AztecAddress } from '@aztec/stdlib/aztec-address';
+import type { ContractDataSource } from '@aztec/stdlib/contract';
 import { SimulationError } from '@aztec/stdlib/errors';
 import { Gas, GasFees } from '@aztec/stdlib/gas';
+import { LogHash } from '@aztec/stdlib/kernel';
+import { ContractClassLogFields } from '@aztec/stdlib/logs';
 import { ProvingRequestType } from '@aztec/stdlib/proofs';
-import { mockTx } from '@aztec/stdlib/testing';
+import { makeContractClassPublic, mockTx } from '@aztec/stdlib/testing';
 import { type MerkleTreeWriteOperations, PublicDataTreeLeaf, PublicDataTreeLeafPreimage } from '@aztec/stdlib/trees';
 import { GlobalVariables, StateReference, Tx, type TxValidator } from '@aztec/stdlib/tx';
 import { getTelemetryClient } from '@aztec/telemetry-client';
@@ -23,7 +28,7 @@ import { PublicProcessor } from './public_processor.js';
 
 describe('public_processor', () => {
   let merkleTree: MockProxy<MerkleTreeWriteOperations>;
-  let contractsDB: MockProxy<PublicContractsDB>;
+  let contractsDB: PublicContractsDB;
   let publicTxSimulator: MockProxy<PublicTxSimulator>;
 
   let mockedEnqueuedCallsResult: PublicTxResult;
@@ -39,9 +44,41 @@ describe('public_processor', () => {
   const mockTxWithPublicCalls = ({ seed = 1, feePayer }: { seed?: number; feePayer?: AztecAddress } = {}) =>
     mockTx(seed, { numberOfNonRevertiblePublicCallRequests: 1, numberOfRevertiblePublicCallRequests: 1, feePayer });
 
+  const mockContractClassForTx = async (tx: Tx, revertible = true) => {
+    const publicContractClass = await makeContractClassPublic(42);
+    const contractClassLogFields = [
+      new Fr(CONTRACT_CLASS_PUBLISHED_MAGIC_VALUE),
+      publicContractClass.id,
+      new Fr(publicContractClass.version),
+      publicContractClass.artifactHash,
+      publicContractClass.privateFunctionsRoot,
+      ...bufferAsFields(
+        publicContractClass.packedBytecode,
+        Math.ceil(publicContractClass.packedBytecode.length / 31) + 1,
+      ),
+    ];
+    const contractAddress = new AztecAddress(new Fr(CONTRACT_CLASS_REGISTRY_CONTRACT_ADDRESS));
+    const emittedLength = contractClassLogFields.length;
+    const logFields = ContractClassLogFields.fromEmittedFields(contractClassLogFields);
+
+    tx.contractClassLogFields.push(logFields);
+
+    const contractClassLogHash = LogHash.from({
+      value: await logFields.hash(),
+      length: emittedLength,
+    }).scope(contractAddress);
+    if (revertible) {
+      tx.data.forPublic!.revertibleAccumulatedData.contractClassLogsHashes[0] = contractClassLogHash;
+    } else {
+      tx.data.forPublic!.nonRevertibleAccumulatedData.contractClassLogsHashes[0] = contractClassLogHash;
+    }
+
+    return publicContractClass.id;
+  };
+
   beforeEach(() => {
     merkleTree = mock<MerkleTreeWriteOperations>();
-    contractsDB = mock<PublicContractsDB>();
+    contractsDB = new PublicContractsDB(mock<ContractDataSource>());
     publicTxSimulator = mock();
 
     const stateReference = StateReference.empty();
@@ -281,5 +318,66 @@ describe('public_processor', () => {
       expect(merkleTree.revertCheckpoint).toHaveBeenCalledTimes(1);
       expect(merkleTree.sequentialInsert).toHaveBeenCalledTimes(0);
     });
+  });
+
+  describe('contract class logs', () => {
+    it.each([
+      [' not', 'revertible'],
+      ['', 'non-revertible'],
+    ])('after a revert, does%s retain contract classes emitted from %s logs', async (_, kind) => {
+      const tx = await mockTxWithPublicCalls();
+
+      const contractClassId = await mockContractClassForTx(tx, kind === 'revertible');
+
+      // Mock the simulator to return a reverted result
+      mockedEnqueuedCallsResult.revertCode = RevertCode.APP_LOGIC_REVERTED;
+      mockedEnqueuedCallsResult.revertReason = new SimulationError('Simulation Failed in app logic', []);
+
+      // Mock the simulator to add contracts to the DB (simulating what the real simulator does)
+      publicTxSimulator.simulate.mockImplementation(async (simulatedTx: Tx) => {
+        // Add contracts from the tx to the contractsDB's revertible and non-revertible tx-level caches
+        await contractsDB.addNewContracts(simulatedTx);
+        return Promise.resolve(mockedEnqueuedCallsResult);
+      });
+
+      const [processed, failed] = await processor.process([tx]);
+
+      // Transaction should be processed but with revert
+      expect(processed.length).toBe(1);
+      expect(failed).toEqual([]);
+
+      // Check whether the contract class is in the DB based on whether it was revertible
+      const contractClass = await contractsDB.getContractClass(contractClassId);
+      // On revert, the public processor only commits non-revertible contracts to the block-level cache
+      // On success, it commits both revertible and non-revertible contracts to the block-level cache
+      if (kind === 'revertible') {
+        expect(contractClass).toBeUndefined();
+      } else {
+        expect(contractClass).toBeDefined();
+      }
+    });
+  });
+
+  // on uncaught error, public processor clears the tx-level cache entirely
+  it('clears the tx-level cache entirely on uncaught error (like SETUP failure)', async function () {
+    const tx = await mockTxWithPublicCalls();
+
+    // we want to confirm that even non-revertibles get cleared
+    const contractClassId = await mockContractClassForTx(tx, /*revertible=*/ false);
+
+    publicTxSimulator.simulate.mockImplementation(async (simulatedTx: Tx) => {
+      await contractsDB.addNewContracts(simulatedTx);
+      throw new Error('Uncaught error');
+    });
+
+    const [processed, failed] = await processor.process([tx]);
+
+    expect(processed).toEqual([]);
+    expect(failed).toEqual([expect.objectContaining({ error: new Error(`Uncaught error`) })]);
+
+    // Check whether the contract class is in the DB based on whether it was revertible
+    const contractClass = await contractsDB.getContractClass(contractClassId);
+    // On uncaught error, the public processor clears the tx-level cache entirely
+    expect(contractClass).toBeUndefined();
   });
 });
